Handle video files or youtube urls in image list
