+++ src/gpt_langchain.py
@@ -71,7 +71,7 @@ from utils import wrapped_partial, EThread, import_matplotlib, sanitize_filename
     get_accordion, have_jq, get_doc, get_source, get_token_count, reverse_ucurve_list, get_size, \
     get_test_name_core, download_simple, have_fiftyone, have_librosa, return_good_url, n_gpus_global, \
     get_accordion_named, hyde_titles, have_cv2, FullSet, create_relative_symlink, split_list, get_gradio_tmp, \
-    merge_dict, get_docs_tokens, markdown_to_html, is_markdown, AsyncNullContext
+    merge_dict, get_docs_tokens, markdown_to_html, is_markdown, AsyncNullContext, url_prefixes_youtube
 from enums import DocumentSubset, no_lora_str, model_token_mapping, source_prefix, source_postfix, non_query_commands, \
     LangChainAction, LangChainMode, DocumentChoice, LangChainTypes, font_size, head_acc, super_source_prefix, \
     super_source_postfix, langchain_modes_intrinsic, get_langchain_prompts, LangChainAgent, docs_joiner_default, \
@@ -4091,191 +4091,6 @@ class Crawler:
                 depth += 1
         self.final_urls = sorted(set(self.urls_to_visit + self.visited_urls))
         return self.final_urls
-
-
-def get_youtube_urls():
-    # https://www.netify.ai/resources/applications/youtube
-    base = ['googlevideo.com',
-            'video.google.com',
-            'video.l.google.com',
-            'wide-youtube.l.google.com',
-            'youtu.be',
-            'youtube.ae',
-            'youtube.al',
-            'youtube.am',
-            'youtube.at',
-            'youtube.az',
-            'youtube.ba',
-            'youtube.be',
-            'youtube.bg',
-            'youtube.bh',
-            'youtube.bo',
-            'youtube.by',
-            'youtube.ca',
-            'youtube.cat',
-            'youtube.ch',
-            'youtube.cl',
-            'youtube.co',
-            'youtube.co.ae',
-            'youtube.co.at',
-            'youtube.co.cr',
-            'youtube.co.hu',
-            'youtube.co.id',
-            'youtube.co.il',
-            'youtube.co.in',
-            'youtube.co.jp',
-            'youtube.co.ke',
-            'youtube.co.kr',
-            'youtube.com',
-            'youtube.co.ma',
-            'youtube.com.ar',
-            'youtube.com.au',
-            'youtube.com.az',
-            'youtube.com.bd',
-            'youtube.com.bh',
-            'youtube.com.bo',
-            'youtube.com.br',
-            'youtube.com.by',
-            'youtube.com.co',
-            'youtube.com.do',
-            'youtube.com.ec',
-            'youtube.com.ee',
-            'youtube.com.eg',
-            'youtube.com.es',
-            'youtube.com.gh',
-            'youtube.com.gr',
-            'youtube.com.gt',
-            'youtube.com.hk',
-            'youtube.com.hn',
-            'youtube.com.hr',
-            'youtube.com.jm',
-            'youtube.com.jo',
-            'youtube.com.kw',
-            'youtube.com.lb',
-            'youtube.com.lv',
-            'youtube.com.ly',
-            'youtube.com.mk',
-            'youtube.com.mt',
-            'youtube.com.mx',
-            'youtube.com.my',
-            'youtube.com.ng',
-            'youtube.com.ni',
-            'youtube.com.om',
-            'youtube.com.pa',
-            'youtube.com.pe',
-            'youtube.com.ph',
-            'youtube.com.pk',
-            'youtube.com.pt',
-            'youtube.com.py',
-            'youtube.com.qa',
-            'youtube.com.ro',
-            'youtube.com.sa',
-            'youtube.com.sg',
-            'youtube.com.sv',
-            'youtube.com.tn',
-            'youtube.com.tr',
-            'youtube.com.tw',
-            'youtube.com.ua',
-            'youtube.com.uy',
-            'youtube.com.ve',
-            'youtube.co.nz',
-            'youtube.co.th',
-            'youtube.co.tz',
-            'youtube.co.ug',
-            'youtube.co.uk',
-            'youtube.co.ve',
-            'youtube.co.za',
-            'youtube.co.zw',
-            'youtube.cr',
-            'youtube.cz',
-            'youtube.de',
-            'youtube.dk',
-            'youtubeeducation.com',
-            'youtube.ee',
-            'youtubeembeddedplayer.googleapis.com',
-            'youtube.es',
-            'youtube.fi',
-            'youtube.fr',
-            'youtube.ge',
-            'youtube.googleapis.com',
-            'youtube.gr',
-            'youtube.gt',
-            'youtube.hk',
-            'youtube.hr',
-            'youtube.hu',
-            'youtube.ie',
-            'youtubei.googleapis.com',
-            'youtube.in',
-            'youtube.iq',
-            'youtube.is',
-            'youtube.it',
-            'youtube.jo',
-            'youtube.jp',
-            'youtubekids.com',
-            'youtube.kr',
-            'youtube.kz',
-            'youtube.la',
-            'youtube.lk',
-            'youtube.lt',
-            'youtube.lu',
-            'youtube.lv',
-            'youtube.ly',
-            'youtube.ma',
-            'youtube.md',
-            'youtube.me',
-            'youtube.mk',
-            'youtube.mn',
-            'youtube.mx',
-            'youtube.my',
-            'youtube.ng',
-            'youtube.ni',
-            'youtube.nl',
-            'youtube.no',
-            'youtube-nocookie.com',
-            'youtube.pa',
-            'youtube.pe',
-            'youtube.ph',
-            'youtube.pk',
-            'youtube.pl',
-            'youtube.pr',
-            'youtube.pt',
-            'youtube.qa',
-            'youtube.ro',
-            'youtube.rs',
-            'youtube.ru',
-            'youtube.sa',
-            'youtube.se',
-            'youtube.sg',
-            'youtube.si',
-            'youtube.sk',
-            'youtube.sn',
-            'youtube.soy',
-            'youtube.sv',
-            'youtube.tn',
-            'youtube.tv',
-            'youtube.ua',
-            'youtube.ug',
-            'youtube-ui.l.google.com',
-            'youtube.uy',
-            'youtube.vn',
-            'yt3.ggpht.com',
-            'yt.be',
-            'ytimg.com',
-            'ytimg.l.google.com',
-            'ytkids.app.goo.gl',
-            'yt-video-upload.l.google.com']
-
-    url_prefixes_youtube1 = []
-    for x in base:
-        url_prefixes_youtube1.extend([
-            # '%s/watch?v=' % x,
-            '%s' % x,
-            # '%s/shorts/' % x,
-        ])
-    return set(url_prefixes_youtube1)
-
-
-url_prefixes_youtube = get_youtube_urls()
 
 
 def file_to_doc(file,
