+++ src/version.py
@@ -1,1 +1,1 @@
-__version__ = "5254d6cf2e55de2862bd29ceaafd07bab10b6b53"
+__version__ = "f2e91553ca6334ff723b1ca00e67f0fe13952969"
